Search Bar: Move the related components to their own dir
Move the Search Bar components into a dedicated directory
