+++ app/src/lib/components/RemoteBranchPreview.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
 	import RemoteCommitList from './RemoteCommitList.svelte';
-	import SearchBar from './SearchBar.svelte';
-	import BranchPreviewHeader from '../branch/BranchPreviewHeader.svelte';
-	import Resizer from '../shared/Resizer.svelte';
-	import ScrollableContainer from '../shared/ScrollableContainer.svelte';
 	import { Project } from '$lib/backend/projects';
+	import BranchPreviewHeader from '$lib/branch/BranchPreviewHeader.svelte';
 	import FileCard from '$lib/file/FileCard.svelte';
+	import SearchBar from '$lib/searchBar/SearchBar.svelte';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
+	import Resizer from '$lib/shared/Resizer.svelte';
+	import ScrollableContainer from '$lib/shared/ScrollableContainer.svelte';
 	import { getRemoteBranchData } from '$lib/stores/remoteBranches';
 	import { getContext, getContextStore, getContextStoreBySymbol } from '$lib/utils/context';
 	import { FileIdSelection } from '$lib/vbranches/fileIdSelection';
