+++ app/src/lib/searchBar/FilterPill.svelte
