+++ app/src/lib/searchBar/FilterPillContainer.svelte
