+++ app/src/routes/[projectId]/base/+page.svelte
@@ -2,8 +2,8 @@
 	import { Project } from '$lib/backend/projects';
 	import BaseBranch from '$lib/components/BaseBranch.svelte';
 	import FullviewLoading from '$lib/components/FullviewLoading.svelte';
-	import SearchBar from '$lib/components/SearchBar.svelte';
 	import FileCard from '$lib/file/FileCard.svelte';
+	import SearchBar from '$lib/searchBar/SearchBar.svelte';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import Resizer from '$lib/shared/Resizer.svelte';
 	import ScrollableContainer from '$lib/shared/ScrollableContainer.svelte';
